refactor(roomList): organizes RooomList componets and subcomponents files
