+++ client/src/components/roomList/RoomList.jsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 import styles from './RoomList.module.css';
-import RoomListPlayers from '../roomListPlayers/RoomListPlayers';
-import RoomPlayerForm from '../roomPlayerForm/RoomPlayerForm';
+import RoomListPlayers from './roomListPlayers/RoomListPlayers.jsx';
+import RoomPlayerForm from './roomPlayerForm/RoomPlayerForm';
 
 function RoomList({ rooms, onJoinRoom }) {
   return (
